refactor: more prototype based to classes
